requirements.txt
fastapi
bson
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
python-jose[cryptography]
passlib
bcrypt
python-multipart
PyJWT
motor
uuid